backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
sqlalchemy==2.0.30
aiosqlite==0.19.0
asyncpg==0.29.0
anthropic==0.18.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.24.1
loguru==0.7.2
fpdf2==2.7.9
